+++ kitty/options/to-c-generated.h
@@ -58,41 +58,15 @@ convert_from_opts_modify_font(PyObject *py_opts, Options *opts) {
 }
 
 static void
-convert_from_python_text_old_gamma(PyObject *val, Options *opts) {
-    opts->text_old_gamma = PyObject_IsTrue(val);
+convert_from_python_text_composition_strategy(PyObject *val, Options *opts) {
+    text_composition_strategy(val, opts);
 }
 
 static void
-convert_from_opts_text_old_gamma(PyObject *py_opts, Options *opts) {
-    PyObject *ret = PyObject_GetAttrString(py_opts, "text_old_gamma");
+convert_from_opts_text_composition_strategy(PyObject *py_opts, Options *opts) {
+    PyObject *ret = PyObject_GetAttrString(py_opts, "text_composition_strategy");
     if (ret == NULL) return;
-    convert_from_python_text_old_gamma(ret, opts);
-    Py_DECREF(ret);
-}
-
-static void
-convert_from_python_text_gamma_adjustment(PyObject *val, Options *opts) {
-    opts->text_gamma_adjustment = PyFloat_AsFloat(val);
-}
-
-static void
-convert_from_opts_text_gamma_adjustment(PyObject *py_opts, Options *opts) {
-    PyObject *ret = PyObject_GetAttrString(py_opts, "text_gamma_adjustment");
-    if (ret == NULL) return;
-    convert_from_python_text_gamma_adjustment(ret, opts);
-    Py_DECREF(ret);
-}
-
-static void
-convert_from_python_text_contrast(PyObject *val, Options *opts) {
-    opts->text_contrast = PyFloat_AsFloat(val);
-}
-
-static void
-convert_from_opts_text_contrast(PyObject *py_opts, Options *opts) {
-    PyObject *ret = PyObject_GetAttrString(py_opts, "text_contrast");
-    if (ret == NULL) return;
-    convert_from_python_text_contrast(ret, opts);
+    convert_from_python_text_composition_strategy(ret, opts);
     Py_DECREF(ret);
 }
 
@@ -1081,11 +1055,7 @@ convert_opts_from_python_opts(PyObject *py_opts, Options *opts) {
     if (PyErr_Occurred()) return false;
     convert_from_opts_modify_font(py_opts, opts);
     if (PyErr_Occurred()) return false;
-    convert_from_opts_text_old_gamma(py_opts, opts);
-    if (PyErr_Occurred()) return false;
-    convert_from_opts_text_gamma_adjustment(py_opts, opts);
-    if (PyErr_Occurred()) return false;
-    convert_from_opts_text_contrast(py_opts, opts);
+    convert_from_opts_text_composition_strategy(py_opts, opts);
     if (PyErr_Occurred()) return false;
     convert_from_opts_cursor_shape(py_opts, opts);
     if (PyErr_Occurred()) return false;